+++ mpf/mpfconfig.yaml
@@ -86,8 +86,7 @@ mpf:
         p3_roc:  mpf.platforms.p3_roc.HardwarePlatform
         pololu_maestro: mpf.platforms.pololu_maestro.HardwarePlatform
         smart_virtual: mpf.platforms.smart_virtual.HardwarePlatform
-        smartmatrix: mpf.platforms.smartmatrix.HardwarePlatform
-        eli_dmd: mpf.platforms.eli_dmd.EliDmd
+        smartmatrix: mpf.platforms.smartmatrix.SmartMatrix
         snux: mpf.platforms.snux.HardwarePlatform
         virtual: mpf.platforms.virtual.HardwarePlatform
