+++ mpf/platforms/smartmatrix.py
@@ -1,29 +1,25 @@
-"""Contains code for an SmartMatrix Shield connected to a Teensy."""
+"""Contains code for SmartMatrix RGB DMD."""
 
 import logging
-import sys
-import threading
-import traceback
-from queue import Queue
-import serial
+
+import asyncio
 from mpf.core.platform import RgbDmdPlatform
 
 
-class HardwarePlatform(RgbDmdPlatform):
+class SmartMatrix(RgbDmdPlatform):
 
-    """SmartMatrix shield via Teensy."""
+    """SmartMatrix RGB DMD."""
 
     def __init__(self, machine):
-        """Initialise smart matrix."""
+        """Initialise RGB DMD."""
         super().__init__(machine)
+        self.features['tickless'] = True
 
         self.log = logging.getLogger('SmartMatrix')
-        self.log.debug("Configuring SmartMatrix hardware interface.")
+        self.log.debug("Configuring SmartMatrix RGB DMD hardware interface.")
 
-        self.queue = None
-        self.serial_port = None
-        self.dmd_thread = None
-        self.update = None
+        self.reader = None
+        self.writer = None
 
         self.config = self.machine.config_validator.validate_config(
             config_spec='smartmatrix',
@@ -36,8 +32,8 @@ class HardwarePlatform(RgbDmdPlatform):
     def stop(self):
         """Stop platform."""
         try:
-            self.log.info("Disconnecting from SmartMatrix hardware...")
-            self.serial_port.close()
+            self.log.info("Disconnecting from SmartMatrix RGB DMD hardware...")
+            self.writer.close()
         except AttributeError:
             pass
 
@@ -47,42 +43,28 @@ class HardwarePlatform(RgbDmdPlatform):
 
     def configure_rgb_dmd(self):
         """Configure rgb dmd."""
-        self.log.info("Connecting to SmartMatrix DMD on %s", self.config['port'])
-        self.serial_port = serial.Serial(port=self.config['port'],
-                                         baudrate=2500000)
-
-        if self.config['use_separate_thread']:
-            self.queue = Queue()
-            self.dmd_thread = threading.Thread(target=self._dmd_sender_thread)
-            self.dmd_thread.daemon = True
-            self.dmd_thread.start()
-            self.update = self._update_separate_thread
-        else:
-            self.update = self._update_non_thread
-
+        self.machine.clock.loop.run_until_complete(self._connect())
         return self
 
-    def _update_non_thread(self, data):
-        try:
-            self.serial_port.write(bytearray([0x01]))
-            self.serial_port.write(bytearray(data))
-        except TypeError:
-            pass
+    @staticmethod
+    def _done(future):
+        """Evaluate result of task.
 
-    def _update_separate_thread(self, data):
-        self.queue.put(bytearray(data))
+        Will raise exceptions from within task.
+        """
+        future.result()
 
-    def _dmd_sender_thread(self):
-        while True:
-            data = self.queue.get()  # this will block
+    @asyncio.coroutine
+    def _connect(self):
+        self.log.info("Connecting to SmartMatrix RGB DMD on %s baud %s", self.config['port'], self.config['baud'])
+        connector = self.machine.clock.open_serial_connection(
+            url=self.config['port'], baudrate=self.config['baud'], limit=0)
+        self.reader, self.writer = yield from connector
 
-            try:
-                self.serial_port.write(bytearray([0x01]))
-                self.serial_port.write(bytearray(data))
-
-            except IOError:
-                exc_type, exc_value, exc_traceback = sys.exc_info()
-                lines = traceback.format_exception(exc_type, exc_value,
-                                                   exc_traceback)
-                msg = ''.join(line for line in lines)
-                self.machine.crash_queue.put(msg)
+    def update(self, data):
+        if self.writer:
+            if self.config['old_cookie']:
+                self.writer.write(bytearray([0x01]))
+            else:
+                self.writer.write(bytearray([0xBA, 0x11, 0x00, 0x03, 0x04, 0x00, 0x00, 0x00]))
+            self.writer.write(bytearray(data))
